Intégré la dernière version du web-service Party 3 qui restreint la plage de validité des numéros de contribuables (conversion long -> int sur les numéros de contribuable).
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@62582 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/web-it/src/test/java/ch/vd/uniregctb/webservice/tiers3/TiersServiceWebCEDITest.java
@@ -333,7 +333,7 @@ public class TiersServiceWebCEDITest extends AbstractTiersServiceWebTest {
 
 		GetPartyRequest params = new GetPartyRequest();
 		params.setLogin(login);
-		params.setPartyNumber(ctbID);
+		params.setPartyNumber((int) ctbID);
 		params.getParts().add(PartyPart.TAX_DECLARATIONS);
 		params.getParts().add(PartyPart.TAX_DECLARATIONS_STATUSES);
 
@@ -371,7 +371,7 @@ public class TiersServiceWebCEDITest extends AbstractTiersServiceWebTest {
 		demande.setReturnDate(dateRetour);
 
 		final OrdinaryTaxDeclarationKey di0 = new OrdinaryTaxDeclarationKey();
-		di0.setTaxpayerNumber(ctbId);
+		di0.setTaxpayerNumber((int) ctbId);
 		di0.setSequenceNumber(numeroSequenceDI);
 		di0.setTaxPeriod(periodeFiscale);
 		demande.setKey(di0);
