+++ unireg/ws/src/main/java/ch/vd/uniregctb/webservices/tiers3/data/RelationBetweenPartiesBuilder.java
@@ -11,7 +11,7 @@ public class RelationBetweenPartiesBuilder {
 		r.setDateFrom(DataHelper.coreToWeb(rapport.getDateDebut()));
 		r.setDateTo(DataHelper.coreToWeb(rapport.getDateFin()));
 		r.setCancellationDate(DataHelper.coreToWeb(rapport.getAnnulationDate()));
-		r.setOtherPartyNumber(autreTiersNumero);
+		r.setOtherPartyNumber(autreTiersNumero.intValue());
 
 		if (rapport instanceof ch.vd.uniregctb.tiers.RapportPrestationImposable) {
 			final ch.vd.uniregctb.tiers.RapportPrestationImposable rpi = (ch.vd.uniregctb.tiers.RapportPrestationImposable) rapport;
