+++ unireg/ws/src/main/java/ch/vd/uniregctb/webservices/tiers3/impl/DataHelper.java
@@ -228,7 +228,7 @@ public class DataHelper {
 		}
 
 		final PartyInfo i = new PartyInfo();
-		i.setNumber(value.getNumero());
+		i.setNumber(value.getNumero().intValue());
 		i.setName1(value.getNom1());
 		i.setName2(value.getNom2());
 		i.setStreet(value.getRue());
